Update NavigationBar.jsx
Replaced the static navigation links with the more dynamic NavigationLinks( { id, role } ).filter( ... ).map( ... ).

Tested it out by removing { id, role } arguments. It works.

NOTE: The .filter is used to show only roles that are relevant to THAT logged in user's role. The descriptions for the restrictions can be seen in NavigationLinks.jsx.