filtered and displayed the shifts signed up.
Replaced shiftsAppliedFor.map( ) with shiftsArray.filter( ).map( ) because, what I want to do is to: 1. FILTER out the shifts in the shiftsArray state that are not inluded in the array of shiftsAppliedFor (the shifts I checked). Then... 2. I want to map it out.

I also changed the result so that the shift details show instead of just the id.