Update App.jsx
Removed console.logs but more importantly, inside the return ( ) => { ... } of useEffect, I added two setStates that are designed to RESET the data back to the original state.

WHY?

Because I found out that, every time I make an adjustment in my code... even a console.log( ), the state would re-render and, if that state is an array, it would add (.push( )) the same info that is already inside.