New Setup.
Removed the Promise.all and replaced with functionDeclarations that I would call inside the useEffectHook: callFunctionDeclarations(functionDeclaratioons).

The reason: Promise.all( ) calls all the promises CONCURRENTLY, which is not what I want. I want the promises to be called one after another so as to not risk the "prv" inside the setState resetting the state to initial b/c (I believe) Promise.all would have all the promises start at the initial state.