Refactored didApplicantApplyForShift function.
This refactoring was based on the fact that the id for shiftsWithApplicants object is now the currentUserID + shiftID to ensure that the same user (currentUserID) cannot apply for the same shift.

A DIFFERENT user (different currentUserID) can apply for the same shift as the prv. user and the shiftsWithApplicants object id would be the new currentUserID concatenated with the same shift ID. This is to ensure uniqueness.