Imported functions.
FetchShiftStatuses which I plan to use to setState for shiftStatuses and TransferApprovedShift, which:

(1) Deletes not only the approved applicant (the applicant the supervisor selected)...

Balso ALL the other applicants who applied and was NOT selected (see this function, deleteApplicantsFromSWA, which is inside TransferApprovedShift).

(2) For the selected (approved) applicant, it moves that selected applicant to http://localhost:3003/shiftsPendingConfirmation, because now, after the supervisor APPROVES the shift, I set this up so that the applicant has to ACKNOWLEDGE that approval (confirm the approved shift):

const AddToShiftsPendingConfirmation = await PostRequestII('http://localhost:3003/shiftsPendingConfirmation', approved_shift_updated, null);

I also called the FetchShiftStatuses function (see FetchHook.js to see the logic of FetchShiftStatuses). The returned shiftStatuses for FetchShiftStatuses is used to setState for shiftStatuses (updates new state).